unify charge point message components
